requirements.txt
torch
torchvision
opencv-python
numpy
scikit-learn
transformers
ultralytics
segment-anything
pycocotools
Pillow